chore: move tp transformers in control-plane-transfomer
