+++ extensions/control-plane/api/management-api/transfer-process-api/build.gradle.kts
@@ -27,6 +27,7 @@ dependencies {
     implementation(project(":core:common:lib:validator-lib"))
     implementation(project(":extensions:common:api:lib:management-api-lib"))
     implementation(project(":extensions:common:http:lib:jersey-providers-lib"))
+    implementation(project(":core:control-plane:control-plane-transform"))
     implementation(libs.jakarta.rsApi)
 
     testImplementation(project(":core:common:lib:transform-lib"))
